requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
flake8==7.1.1
flake8-docstrings==1.7.0
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
packaging==25.0
pluggy==1.6.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
pyTelegramBotAPI==4.22.1
pytest==8.3.3
pytest-timeout==2.3.1
python-dotenv==1.0.1
requests==2.32.3
snowballstemmer==3.0.1
urllib3==2.4.0
